FAT: Add CONFIG_VFAT_FS_NO_DUALNAMES option
This patch adds a new config option VFAT_FS_NO_DUALNAMES. When
VFAT_FS_NO_DUALNAMES is enabled, VFAT will never create a file with
both a long and a short filename. If a long filename is needed, then
dummy bytes are stored in the 8.3 entry, and the 8.3 entry is marked
as having no short name using the FAT_NO_83NAME flag.

Signed-off-by: Andrew Tridgell <tridge@samba.org>