FogBugz #184646: Turn on all peripheral clocks for a system reboot
commit f78b6b5c924fbe6e740117a0495495d4d1f65911 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

When doing a software reboot, all peripheral clocks must get turned on for the
L3 interconnect to work.

This code is needed when doing a "reboot" from user-space and a peripheral
clock as been gated off. Why would a peripheral clock get gated? An example
use case would be a .ko that gets insmod and rmmod during runtime. The insmod
would turn on the IP's clock, and the rmmod would turn off the IP's clock.
Doing a "reboot" would cause the system to hang.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>

Conflicts:

	arch/arm/mach-socfpga/core.h
Signed-off-by: Liwei Song <liwei.song@windriver.com>