v4l: xilinx: dma: Remove colorspace check in xvip_dma_verify_format
commit ba3ea403345c5a2953ce4476d2b78f482c40af33 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

In current implementation driver only checks the colorspace
between the last subdev in the pipeline and the connected video node,
the pipeline could be configured with wrong colorspace information
until the very end. It thus makes little sense to check the
colorspace only at the video node. So check can be dropped until
we find a better solution to carry colorspace information
through pipelines and to userspace.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Acked-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: waiting (media)
Link: https://www.spinics.net/lists/linux-media/msg133380.html
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>