platform: tegra: Add support for mask with 1s
commit c8357acfe0f3c066473f34c480247941a36829f7 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git jetson_36.3

BugLink: https://bugs.launchpad.net/bugs/2028400

Currently, the value of prod setting are applied if mask
bit is 0 for given value. This is non-traditional style of
defining mask.

Add support a traditional way of defining mask like if mask bits
are 1s then only apply the value of those bits.

	reg = (reg & ~mask) | (reg & mask);

This mechanism is selected via DT property to have backward
compatibility. The DT property is "mask-one-style" which can be
added under prod-setting node.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>