crypto: caam - add support for MOVEB command
commit  e80af49455e6a3c8cd93f88719d083e87e90ecb8 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

CHAs of SEC work natively in BE mode. When moving
data to the alignment blocks, swapping is needed
for LE platforms. This is done by means of the MOVEB
command. This patch adds support
to DCL for this command.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Signed-off-by: Radu Alexe <radu.alexe@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>