media: xilinx: dma: Enable use of Framebuffer Write IP in V4L2 Driver
commit f9bb5365c8e713b2abc23da48d5c402f104fb137 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

The Xilinx Video Framebuffer driver is video 'format aware' and
requires clients to pass video format information as part of DMA
programming. As the DMA Engine interface does not support passing
this information, a configuration object is passed through the
dma channel's private data to the Framebuffer Driver with this
information.

Signed-off-by: Jeffrey Mouroux <jeff.mouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>