media: xilinx: dma: Enable use of Framebuffer Write IP in V4L2 Driver
commit 76a0a11c67def390ef265fb1baf28420eb2387a7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The Xilinx Video Framebuffer driver is video 'format aware' and
requires clients to pass video format information as part of DMA
programming. As the DMA Engine interface does not support passing
this information, a configuration object is passed through the
dma channel's private data to the Framebuffer Driver with this
information.

Signed-off-by: Jeffrey Mouroux <jeff.mouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>