gpu: imx: dpu: extdst: Add extdst_pixengcfg_syncmode_master() helper
commit  1047c0120a080b4e9f1bd90c31ec613051a28731 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

This patch adds extdst_pixengcfg_syncmode_master() helper support
so that the callers may control if a extdst is master or slave
when it works in sync mode.  The bit16 of extdst's PIXENGCFG_STATIC
register controls this and it's a part of sync mode fixup logic.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>