v4l: dma: Use media graph traversal to stream on/off subdevs
commit a518ce7081cd59b54dc8541c0d98df0069eab709 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

This patch uses media framework specific implementation to parse
all the subdevs in the graph in order to stream on/off.

Depth first search algorithm is employed to traverse the nodes
and leaf node of the graph will stream on/off first and so on.

Case 1:
cam -> mipi -> scaler -> dma
The order of stream on is cam, mipi and scaler.
The order of stream off is cam, mipi and scaler.

Case 2:
cam  |      | dma
       mipi
cam2 |      | dma2
The order of stream on is cam, mipi and cam2.

Case 3 (mem2mem):
dma -> scaler -> colconv -> dma2
The order of stream on is colconv, scaler. In mem2mem graph will
traverse same subdevs twice but won't stream-on if it is already
streaming.

Signed-off-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
[michals: v6.6 Squashed with
xilinx: v4l: dma: Remove unused function definition
v6.12 update
xilinx: v4l: dma: In case of error xvip_pipeline_start_stop should return gracefully
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>