staging: fsl_ppfe/eth: add fixed-link support
commit  d9c9cdd78712cae5eb0728b1096125e932ebf9e5 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

In cases where MAC is not connected to a normal MDIO-managed PHY
device, and instead to a switch, it is configured as a "fixed-link".
Code to handle this scenario is added here.

phy_node in the dtb is checked to identify a fixed-link.
On identification of a fixed-link, it is registered and connected.

Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>