ahci: qoriq: workaround for errata A-379364 on lx2160a
commit  e5797c368bfe2dafbade1e87cd292f98182dee03 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

There is a erratum on lx2160a which is: "SATA link is
going down sometime during sata initialization"
The workaround for it is to reset the lane. This patch
implements this workaround.
This erratum only exists on lx2160 Rev1, will be addressed
on Rev2 and later.

Signed-off-by: Peng Ma <peng.ma@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>