MLK-10600-1: mxc: sim: Add SIM IP driver support
commit 147fb40ccb4d4ab4d19e8b7ece25f3ceda4340f3 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

This driver is based on the current code which runs the the EMV test on the i.MX258 platform.
Since there are still many cases that can't pass on the i.MX258 and i.MX7d platform. The
driver will need to be improved after per-test work. Just check in as a base code. There
would be definitly some timing improvement work to do in the future.

Signed-off-by: Luwei Zhou <b45643@freescale.com>
Signed-off-by: Gao Pan <b45643@freescale.com>
(cherry picked from 3ac1ad5b2a68ecb052ccacca4ac7459ead04415e)
[ Aisheng: squash legacy patches ]
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
[Xiaolei resolved the backport conflict.]
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>