HSD #1806784746: EDAC/altera: Warm Reset option S10 DBE
commit e3d7e62874653ebc3c08bdfc2fa307a17f3115b7 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

The Stratix10 peripheral FIFO memories can recover from double
bit errors with a warm reset instead of a cold reset.
Add the option of a warm reset for peripheral (USB, Ethernet)
memories.

CPU memories such as SDRAM and OCRAM require a cold reset for
DBEs.
Filter on whether the error is a SDRAM/OCRAM or a peripheral
FIFO memory to determine which reset to use when the warm
reset option is configured.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>