MLK-22158 drm/imx: lcdif: force 'mode_changed' for fb pix fmt change
commit  7bd57d68588fe5e97ae1318396cbd0910899406e from
https://github.com/nxp-imx/linux-imx lf-6.18.y

According to the LCDIF function definition, the pixel format
configuration should not be done during LCDIF is running. So
if finding pixel format change is requested, the 'mode_changed'
needs to be set to make sure the commit become to be a full
modeset.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>