arm: imx: rename the compatible node of lpm mu driver
commit  062918acdfcaec4f668c4276c0338953b7fd1411 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Because that the mailbox MU driver is used in i.MX RPMSG implementation.
There is a confliction between this MU driver and the mailbox MU driver.

To back-compaible with LPM of iMX6SX, iMX7D and iMX7ULP.
Rename the compatible node of this MU driver.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>