drm/imx: dpu: crtc: Do not send out bogus vblank event in ->atomic_disable()
commit  4a82f24401cbf37ac4b6390aa21f360f803fc2fa from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When a full modeset is needed, the CRTC could be totally disabled or
enabled/re-enabled after the modeset.  If it's re-enabled, a vblank
event would be sent during the CRTC enablement procedure.  So, a bogus
event should be killed in the ->atomic_disable() callback.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>