soc: fsl: dpio: fix qbman alignment error in the virtualization context
commit  441d6a57a4bce92c7c93a8de0f14365e65c44de7 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When running as a guest, under KVM, the CENA region is mapped
as device memory, so uncacheable. When the memory is mapped
as device memory, the unaligned accesses are not allowed.
Memcpy is optimized to transfer 8 bytes at a time regardless
of the start address and might cause alignment issues.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>