of: configfs: Fix error handling from of_overlay_create()
commit 6ff7487238738f05cae514022e153f86b5de1e16 from
https://github.com/linux4microchip/linux.git

of_overlay_create() is returning the id or the created overlay,
or a negative error number.
It means when the first overlay is applied, id 0 is returned.
When the second overlay is applied, id 1 is returned which
is not captured by error condition in create_overlay() because it is
valid return value but function itself was returning this directly
back to cfs_overlay_item_path_store() where
if (err != 0) condition traps it and return to out_err label
and return id value as number of chars which was handled by interface.

This patch is fixing error value handling in create_overlay() to return
error value directly when error is handled and also return 0 instead of
overlay id.

The second change is in error path cfs_overlay_item_path_store()
where number of written bytes is return not to call write
again.

For example behaviour with Busybox echo:
write(1, "jdtmp3.dtbo", 11)             = 1
write(1, "dtmp3.dtbo", 10)              = -1 ENOENT (No such
file or directory)
write(2, "echo: write error: No such file "..., 45echo: write
error: No such file or directory) = 45

Signed-off-by: Michal Simek <monstr@monstr.eu>
State: not-upstreamable
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>