MLK-23822: soc: imx: secvio: Fix boot message when nvmem not initialised
commit  f86b36f56244f41d8c1ed2dee11df9d8776cd55e from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When the nvmem subsystem is not initialised at boot, the probe
will fail and an error message will be displayed.
In this case the message should not be printed as the driver will
be probed later.

This patch checks the error code from nvmem before printing the
message.

It also fixes the cleaning path as the driver was not exiting
properly.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>