tty: serial: linflexuart: Resume DMA transfer after polling mode finished
commit 5bb207d8fc1aac2c397aba60f1fdb04232c4b3ac from
https://github.com/nxp-auto-linux/linux

When there are both kernel and userspace output, the
last one can't be seen after a while. The userspace console hangs until
the user wakes up it by sending a character.
The fix is to resume the tty output and the DMA transfer after polling
mode output is finished.

Issue: ALB-5329
Issue: ALB-9101
Signed-off-by: Js Ha <js.ha@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
[Zhantao: using tport->xmit_buf to replace the original xmit settings]
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>