LF-2321-10: drm/mxsfb: Add support for horizontal stride
commit  83a0efdda3cc3644fd3406382e13d2fed3ebc52e from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Besides the eLCDIF block, there is another IP block, used in the
past for EPDC panels. Since the iMX.8mq doesn't have an EPDC connector,
this block is not documented, but we can use it to do additional
operations on the frame buffer.
In this case, we can use the pigeon registers from this IP block in
order to do horizontal crop on the frame buffer processed by the
eLCDIF block.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
[ Liu Ying: Improve the way to get new plane state in primary's ->atomic_update() ]
[ Liu Ying: Fix conflicts to upgrade to next-20220527 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20220912 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>