xhci: quirks: add link TRB quirk for VL805
commit  15667aeb94e12324b405c53b8a69c14318a3dabe from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The VL805 controller can't cope with the TR Dequeue Pointer for an endpoint
being set to a Link TRB. The hardware-maintained endpoint context ends up
stuck at the address of the Link TRB, leading to erroneous ring expansion
events whenever the enqueue pointer wraps to the dequeue position.

If the search for the end of the current TD and ring cycle state lands on
a Link TRB, move to the next segment.

Link: https://github.com/raspberrypi/linux/issues/3919

[6.5.y Fixup - move downstream quirk bits further along]

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>