spi: spi-fsl-dspi: Use DMA for S32CC controller in target mode
commit 06504e712dde5de7aeadba58b61aa96fc6a6ddb3 from
https://github.com/nxp-auto-linux/linux

Switch to DMA for target mode otherwise the controller is too slow to
feed TX FIFO and UNDERFLOW occurs frequently.
DMA can work only with 8 and 16 bits per word. 32bits per word
is not supported, this is a hardware limitation, so we
keep the controller mode in TCFQ mode.

Issue: ALB-1129
Issue: ALB-10889
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>