i2c: designware: Add SMBUS quick command support
commit 9b879b6a24f758a2c251ccd64b114fdc12c51dd4 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The SMBUS emulation code turns an SMBUS quick command into a zero-
length read. This controller can't do zero length accesses, but it
can do quick commands, so reverse the emulation. The alternative
would be to properly implement the SMBUS support but that is a lot
more work, and unnecessary just to get i2cdetect working.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>