LF-3024: drm/imx/dcss: allow primary plane to be less than crtc size
commit  be23374e0138ec03de1268dde7f98ad8d2e9841a from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Currently, the primary plane has to cover the entire CRTC area. But, there are
situations when we want the primary plane to be smaller, to save bandwidth.
Since the HW allows it, activate it in the driver as well.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Acked-by: Mirela Rabulea <mirela.rabulea@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20210514 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>