LF-3299-1 mfd: add core driver for FP9931 PMIC
commit  01379745e930bab17f14a4dbcbfc156ec5ccee22 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

FP9931 is a PMIC IC which comes from Fitipower Corp. It can
be used to for voltages power on or down sequence and timing
configuration for different e-ink panels, and also it has a
thermistor for temperature detection. Add this core driver
to handle regulators device and hwmon device.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: G.n. Zhou <guoniu.zhou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>