dt-bindings: spi: dspi: Add S32CC support
commit 0a7bc1528180e118e7b8bf730e2aa88a13e00464 from
https://github.com/nxp-auto-linux/linux

Including documenation for SPI slave support.
spi-slave property should be used for slave mode only.

The spi controller node in dts cand define both master
and slave pinctrl. The selection between them will be done based
on pinctrl-names. The default pinctrl will be loaded first
and will be used by the master. If the controller is configured
as slave (spi-slave property is added in the dts node), the
driver will look for the "slave" pinctrl and apply it if
found.

Issue: ALB-1129
Issue: ALB-5969
Issue: ALB-10729
Issue: ALB-10889
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>