tty: serial: linflexuart: Fix console freeze
commit 4e4796fcd4ff70f5e7202f1f197c1afca78392f6 from
https://github.com/nxp-auto-linux/linux

According to S32G2_RM_Rev2_DraftF, in Suspend an active DMA channel,
"To suspend an active DMA channel:
1. Stop the DMA service request at the peripheral first. Confirm it has
been disabled by reading back the appropriate register in the peripheral.
2. Check the DMA's Hardware Request Status (HRS) to ensure there is no
service request to the DMA channel being suspended. Then disable the
hardware service request by clearing the ERQ field to 0 on the
appropriate DMA channel."

So we should stop the DMA service request at the peripheral first, which
is achived by clearing DMATXE/DMARXE, before terminating the DMA
transaction.

The issue was seen when a user-space process is trying to print something
and it is interrupted by the kernel after the eDMA transaction was fired
and before any characters are copied by the eDMA in BDRL.

Issue: ALB-5882
Issue: ALB-9101
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>