LF-3319 net: stmmac: enable rx_clk when killing the vlan
commit  08dfc022e2025aa42488d4acc987c3b4d73a00dd from
https://github.com/nxp-imx/linux-imx lf-6.18.y

when killing the vlan, the rx_clk from the phy need to be enabled.
The patch checks the phy's status before killing the vlan. If the phy
is stopped, enable it, remove the vlan, and then stop the phy.

Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>