staging/bcm2835-camera: Add support for DMABUFs
commit  d43e0c36a071398244cbd3e7368452bf3a4e973e from
https://github.com/raspberrypi/linux.git rpi-6.18.y

DMABUFs are all handled by videobuf2, so there is no reason not
to enable support for them.

Note that this driver is still using the vmalloc allocator, so
the buffers it allocates will not be compatible with the codec
or ISP driver that require contiguous buffers. However this
driver should be able to import the buffers allocated by them.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>