arm_mpam: Add a helper to touch an MSC from any CPU
commit b386b05d96824faba4d122601a75d28b2874a2de from
git@git.assembla.com:cavium/WindRiver.linux.git

Resetting RIS entries from the cpuhp callback is easy as the
callback occurs on the correct CPU. This won't be true for any other
caller that wants to reset or configure an MSC.

Add a helper that schedules the provided function if necessary.
Prevent the cpuhp callbacks from changing the MSC state by taking the
cpuhp lock.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>