ACPI / PPTT: Add a helper to fill a cpumask from a cache_id
commit a968c5643f088437b086c544cba2aa5a32c65972 from
git@git.assembla.com:cavium/WindRiver.linux.git

MPAM identifies CPUs by the cache_id in the PPTT cache structure.

The driver needs to know which CPUs are associated with the cache,
the CPUs may not all be online, so cacheinfo does not have the
information.

Add a helper to pull this information out of the PPTT.

CC: Rohit Mathew <Rohit.Mathew@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Jinchao Li <jinchao.li@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>