fpga: bridge: incorrect set to clear freeze_illegal_request register
commit 352c0c25f4648536b09df83669131e9cceb0de2b from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

A Partial Region Controller can be connected to one or more
Freeze Bridge. Each Freeze Bridge has an illegal_request
bit represented in the freeze_illegal_request register.
Thus, instead of just set to clear the illegal_request bit
for first Freeze Bridge, we need to ensure the set to clear
action is applied to which ever Freeze Bridge that has
occurrence of illegal request.

Signed-off-by: Chew, Chiau Ee <chiau.ee.chew@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>