dt-bindings: media: xilinx: Add xlnx,atomic_streamon documentation
commit 108da1820453545db01586632e036965870441d3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

This property allows the media pipeline connected to this video device
to start streaming irrespective of other video devices not streaming in
the same media device. This is for the scenario where some of the
upstream entities are being shared with multiple video devices
(dma engines) in the same media pipeline.

Signed-off-by: Kevin Liu <wentao.liu@xilinx.com>
Reviewed-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Vishal Sagar <vishal.sagar@xilinx.com>
Acked-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>