dt-bindings: media: xilinx: Add xlnx,atomic_streamon documentation
commit 72378a44e982b86acabdfb8c57ee1d3b6f12859c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

This property allows the media pipeline connected to this video device
to start streaming irrespective of other video devices not streaming in
the same media device. This is for the scenario where some of the
upstream entities are being shared with multiple video devices
(dma engines) in the same media pipeline.

Signed-off-by: Kevin Liu <wentao.liu@xilinx.com>
Reviewed-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Vishal Sagar <vishal.sagar@xilinx.com>
Acked-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>