mailbox: llce: Use 16 channels for logger
commit ae1de84cf4c4574e4ba7a63c2ea9ff0f7ff322e8 from
https://github.com/nxp-auto-linux/linux

This commit prepares the ground for LLCE logger driver which
will use SocketCAN interfaces to send the frames to user space.
For this purpose, the logger will use a channel for each interface
and NAPI specific commands (pop, release, enable / disable notifications)
to control the received data. It inherits all operations already
performed by LLCE CAN on RX except the local storage of a frame.

A local buffer for each interface was needed because all LLCE Logger
interfaces are receiving the frames over one single interrupt line and
the ID of the interface can be only identified after the frame is popped.

Issue: ALB-7505
Issue: ALB-9819
Issue: ALB-10882
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>