arm_mpam: Add debugfs entries to show the MSC/RIS the driver discovered
commit 2bd9b5b1ecc7ddaa2c62eaa8537cdec78f62c7aa from
git@git.assembla.com:cavium/WindRiver.linux.git

Not all of MPAM is visible through the resctrl user-space interface.
To make it easy to debug why certain devices were not exposed through
resctrl, allow the properties of the devices to be read through debugfs.

This adds an mpam directory to debugfs, and exposes the devices as well
as the hierarchy that was built.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>