arm_mpam: Add debugfs entries to show the MSC/RIS the driver discovered
commit f4b23db97fe1d557132e69776fba56836d1b4aeb from
git@git.assembla.com:cavium/WindRiver.linux.git

Not all of MPAM is visible through the resctrl user-space interface.
To make it easy to debug why certain devices were not exposed through
resctrl, allow the properties of the devices to be read through debugfs.

This adds an mpam directory to debugfs, and exposes the devices as well
as the hierarchy that was built.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>