arm_mpam: Track bandwidth counter state for overflow and power management
commit 071d940ad00a9938b00dd9cf1817ee9a97934037 from
git@git.assembla.com:cavium/WindRiver.linux.git

Bandwidth counters need to run continuously to correctly reflect the
bandwidth.

The value read may be lower than the previous value read in the case
of overflow and when the hardware is reset due to CPU hotplug.

Add struct mbwu_state to track the bandwidth counter to allow overflow
and power management to be handled.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>