HSD #22013011814-2: crypto: intel_fcs: Remove dev_err from dev/hwrng read function
commit b34e5840327a0f9706f9109dceb1b557c8d14d61 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

Update low level fcs_rng_read function to don't call dev_err if failed to
read random number from SDM to avoid fcs driver error. This function is
only supported in N5X. For non-N5X family, it is expected to be failed in
fcs driver. So, only return error code in fcs_rng_read function and leave
it to dev/hwrng to report the error.

Signed-off-by: Siew Chin Lim <elly.siew.chin.lim@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>