net: stmmac: Resolve poor line rate after switching from TSO off to TSO on
commit 9bce22d99f516cbba3d680fbee7d3a8fd5c17cfa from
https://github.com/intel/linux-intel-lts.git

Clear mss in TDES and call stmmac_enable_tso() to indicate
a new TSO transmission when it is enabled from TSO off using
ethtool command.

TSO Driver is disable when 'priv->tso = 0' in this same function.
The reason this is put as part of fix_features rather than set_features is
because the commit f748be531d70("stmmac: support new GMAC4") has
already introduced the following codes in fix_features:-

+	/* Disable tso if asked by ethtool */
+	if ((priv->plat->tso_en) && (priv->dma_cap.tsoen)) {
+		if (features & NETIF_F_TSO)
+			priv->tso = true;
+		else
+			priv->tso = false;
+	}

Fixes: f748be531d70 ("stmmac: support new GMAC4")
Signed-off-by: Ling Pei Lee <pei.lee.ling@intel.com>
Signed-off-by: Looi Hong Aun <hong.aun.looi@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>