iommu/arm-smmu-v3: Add mpam helpers to query and set state
commit 4ae04a49a4446df00f3f8a9fcdfde957d71a82ce from
git@git.assembla.com:cavium/WindRiver.linux.git

To allow an iommu_group to be moved between resctrl groups as if it
were a CPU thread, the mpam driver needs to be able to set the partid
and pmg for the iommu_group.

Use the properties in the STE, as these only apply to one stream.

The MPAM driver also needs to know the maximum partid and pmg
values that the SMMU can generate. This allows it to determine
the system-wide common supported range of values. Add a helper
to return this id register.

Tested-by: Amit Singh Tomar <amitsinght@marvell.com>
Signed-off-by: James Morse <james.morse@arm.com>
[Bo: update arm_smmu_group_get_mpam() and arm_smmu_group_set_mpam()
to use arm_smmu_master_domain due to changes in commit ad10dce61303
("iommu/arm-smmu-v3: Make smmu_domain->devices into an allocated list")]
[Bo: update arm_smmu_group_set_mpam() to use struct arm_smmu_ste per
57b89048874c ("iommu/arm-smmu-v3: Add a type for the STE")]
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>