kobject: Add kset_get_next_obj() to allow a kset to be walked
commit a2eb1a1945d11a82083f00975b5c456637322ef7 from
git@git.assembla.com:cavium/WindRiver.linux.git

To expose iommu_groups via the resctrl filesystem, the resctrl driver
needs to be able to walk the list of iommu_groups. These are exposed
via sysfs as a kset.

Add kset_get_next_obj() to allow resctrl to walk the kobjects in the
kset.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>