pcie: e1000: Fix crash with surprise hot-unplug when eth1 up
commit e1ab42b23d765c0f39a8772917786758897a8abf from
https://github.com/nxp-auto-linux/linux

When the network interface corresponding to e1000 NIC is up and
e1000 is surprise hot removed the 'e1000_close' and
'e1000_vlan_rx_kill_vid' routines are trying to read/write
registers from e1000 NIC which leads to a Kernel crash.

As part of this commit, we guard these sections using
'pci_channel_offline' call in order to determine if the
e1000 NIC is present or not in the PCIe slot.

Issue: ALB-7899
Issue: ALB-9108
Issue: ALB-10885
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>