cacheinfo: Expose the code to generate a cache-id from a device_node
commit 8b654533c9ceec765d9f937993fc9ba8a9fc5231 from
git@git.assembla.com:cavium/WindRiver.linux.git

The MPAM driver identifies caches by id for use with resctrl. It
needs to know the cache-id when probeing, but the value isn't set
in cacheinfo until device_initcall().

Expose the code that generates the cache-id. The parts of the MPAM
driver that run early can use this to set up the resctrl structures.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>