LF-4836 arm64: imx8mx vpu: make sure output streamon cmd is seet to daemon after buffer is queued
commit  847f27112b0552ba8f20698532ac64c1c6f9a2c6 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

the cmd V4L2_DAEMON_VIDIOC_STREAMON_OUTPUT should be sent to daemon
when the minimum needed buffers are queued.
so we should make sure that when buffers are queued, we send it to daemon.
These is a bug that driver may not send it to daemon due to some statue
issue.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>