arm_mpam: Allow the maximum partid to be overridden from the command line
commit ba8e6ebe4d1f8964f70c166fc66dccd9a7d0bd21 from
git@git.assembla.com:cavium/WindRiver.linux.git

MPAMs bandwidth monitors are only available via resctrl if there are
enough monitors for each combination of partid and pmg to have one.

As it is unlikely anyone built that many monitors, allow the
maximum partid the system will use to be set from the kernel
command-line.

With this, it should be possible for bandwidth monitors to be
enabled by reducing the number of partid in use.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>