MLK-20697-1: drm/panel: panel-raydium-rm67191: Add MIPI_DSI_MODE_NO_EOT_PACKET flag
commit  bc1fda40901eb54bae5f0a07819335fc6aa109c5 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

This flag was automatically used in the DSI driver, making this panel
work, but the way that flag was used was wrong. Since, fixing the use of
this flag in the DSI driver now breaks this panel, just add it here.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>