Revert "drivers: bus: simple-pm-bus: Add support for probing simple bus only devices"
commit  4a8a5442231cf253618ffaa8883328871af8f138 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

This reverts commit 98e96cf80045a383fcc47c58dd4e87b3ae587b3e.

The following commit introduced in v5.15-rc6 breaks i.MX8QM/QXP
suspend/resume. The issue was reported in community [1] and still under
discussion and not conclusion on the proper fix so far.

This patch temporarily revert it in order to make the release work.
i.MX seems not use .sync_state() mentioned in the original commit,
suppose it may be safe as we also havn't used it before.

98e96cf80045 ("drivers: bus: simple-pm-bus: Add support for probing simple bus only devices")

1. https://lore.kernel.org/all/YYu4EglV7SBZU2Iy@ryzen/

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>