soc: xilinx: vcu: Export functions to obtain logicore settings
commit 066ddf8b39698436956c0fba04b9dc59c6a80f6f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

In mainline linux, xlnx vcu logicore settings can only be obtained using
the syscon framework. This is not backwards compatible with the allegro
modules: https://gitenterprise.xilinx.com/xilinx-vcu/vcu-modules. For
example, if the dts defines a logicore register bank instead of using
the syscon framework, the allegro modules will not be able to obtain the
logicore settings. These settings are required for initialization
parameters.

So this patch exports functions to obtain the xlnx vcu logicore settings
regardless if the syscon framework or logicore register bank is used.

Signed-off-by: Dylan Yip <dylan.yip@xilinx.com>
Signed-off-by: Rohit Visavalia <rohit.visavalia@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>