mtd: rawnand: denali_dt: start denali nand driver a little later
After supporting nor flash device for Agilex NAND booting mode,
there will be 2 kinds of flash initialized during kernel boots
up. The mtd0 device node is assigned to NAND flash that is
initialized before nor flash. But for the RSU feature, the
userspac app only thinks the mtd0 is the first partition of
nor flash. So, in order to compatible with it, adjust the NAND
flash driver is loaded in a little later with late_initcall(),
so that mtd0 is assigned to nor flash device.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>