LF-5793 arm64: dts: imx8mq-evk: config the pull up for the goodix touch irq pad
commit  f0f422372d9972bc0c7c017f1e75a402a0dedd2a from
https://github.com/nxp-imx/linux-imx lf-6.18.y

For imx8mq-evk board which connect RM67199 panel, the board design is not correct
for the touch irq pad. The pad on the SoC side is 3.3v, but the pad on the panel
side is 1.8v. Logically, on imx8mq-evk board, need one 3.3v to 1.8v level switch,
but the board lack this. This involve touch interrupt storm when we probe the
touch driver. So add the irq pad pull up to workaround this issue.

Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>