drivers/perf: nxp-s32-ddr-perf: Handle extra CP parameter
Added the handling of extra CP parameter for some perf events.
The value of CP is set via the "counter_cp" attribute passed
along the event name with the "-e" parameter for `perf stat`.

For example, setting the "counter_cp" parameter for the
"write-queue-depth" event in a measurement, is done as following:

The command above will measure the time (in cycles) when the write
queue depth is equal to 0x8.

The only events that support the CP parameter are:
	- selfresh (bits [1:0])
		- CP == 0b00 -> Time not in self-refresh
		- CP == 0b10 -> Time in self-refresh caused by SW/HW
		- CP == 0b11 -> Time in automatic self-refresh
	- write-queue-depth (bits [3:0])
		- CP value -> time measured while WR Queue Depth == CP
	- lp-read-credit-cnt (bits [6:0])
		- CP value -> time measured while LPR CAM entries == CP
	- hp-read-credit-cnt (bits [6:0])
		- CP value -> time measured while HPR CAM entries == CP
	- write-credit-cnt (bits [6:0])
		- CP value -> time measured while WR CAM entries == CP

Issue: ALB-8859
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Adrian Zaharia <Adrian.Zaharia@windriver.com>