octeontx2-bphy-netdev: cnf10k: Fix ptp hardware clock counter conversion
commit c6410f68907ec33cbc2e5fa4dd499af86a705995 from
git@git.assembla.com:cavium/WindRiver.linux.git

Since the RX GPINT is common for all LMAC's per RFOE MHAB, the BPHY
netdev rx packet handler uses  netdev_priv structure of first valid
LMAC until the point where the code reads the PSW and get's the actual
lmacid. From that point onwards, the code uses the priv structure of
the actual lmacid on which the packet is received. There is an issue
in the code where the PTP HW clock counter conversion to logical PTP
clock (SW timecounter) always uses netdev_priv structure of lmacid 0,
which is causing the slave not to synchronize with master because
on the TX side, the code uses proper lmacid for the conversion and
there is always a difference between TX/RX timestamps presented to
ptp4l application. This patch fixes the issue.

Change-Id: I09a5601f26fcc60b9b0d49c5564ab3d10f46723e
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/79420
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>