gpio: siul2-s32cc: replace `regmap_siul2_accessible` with `regmap_access_tables`
commit eb22474d004f86aa3cae39e3abb275dca91db62a from
https://github.com/nxp-auto-linux/linux

The `regmap_siul2_accessible` function checked the following:
- if a register was in the good range [start_pad, end_pad]
- if the first GPIO corresponding to the reg was part of the GPIO bank

The first check could be replaced with `regmap_access_tables` for each
SIUL2 module. The second check was incorrect because not all bits of a
pad are accessible (e.g. PGPDO6). For S32R45, the PGPDO6 bits 15-10
correspond to GPIOs in SIUL20 and bits 9-0 to GPIOs in SIUL21 (granted
that the PGPDO6 offset is the same but the base address differs between
SIUL2 instances). Still, this condition is impossible to check because
the `regmap_siul2_accessible` doesn't know the value to be written.

Issue: ALB-8967
Issue: ALB-10876
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Andra-Teodora Ilie <andra.ilie@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>