v4l: xilinx: sdirxss: No payload handling for SDI 3G Mode
commit 3ab8b518ceb40809e4ad76e0b4a8be6f9d921797 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

This patch handles the video locking for SDI stream in 3G mode, even if the
SDI stream doesn't contain ST352 payloads. This is handled by controlling
the VPID bit in Module ctrl register and updating the v4l2 framework with
proper format details based on the stream information.

Signed-off-by: Ramanjaneyulu Uppari <ramanjaneyulu.uppari@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>