fsl_qbman: ceetm: errata A-010383 workaround for LNI shapers
commit  9a4ef02c7f96fa734f8d302131f034eebd8001be from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When a frame is dequeued by the CEETM, the effective packet size is a
function of the actual packet size, the configured Overhead Accounting
Length (OAL), and the configured Minimum Packet Size (MPS). This
effective packet size is used to update the CEETM shaper credits. This
MPS value should be applied to the actual packet size before the OAL
value is applied. However, the OAL value is applied to the actual packet
size first. Then, the MPS is applied. This results in an incorrect value
being subtracted from the CEETM credit for that particular Logical
Network Interface (LNI) shaper.

To avoid an erroneous shaper credit count, do not configure both OAL and
MPS on an LNI shaper.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>