s32cc: rtc: Save time before system suspend
commit c122e89cec1e68e968e06a554b5c818939f33e56 from
https://github.com/nxp-auto-linux/linux

After suspend/resume we reinitialize the RTC, which causes a subsequent
rtc_read_time() to fail because the hardware counter is now behind what
we had saved in priv->base.
Save the time just before suspend and reapply at resume to fix hwclock
in combination with suspend-to-ram.
Also take into account the amount of time (in seconds) that we are due
to sleep so that it gets reflected on resume.

Issue: ALB-6347
Issue: ALB-7084
Issue: ALB-9114
Issue: ALB-10867
Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>