arm64: dts: ls1028a-rdb-dpdk: remove pause property
commit  ff15a050731caa2b60ddfb6e2b08a84c14b38b82 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

DPDK is not capable to handle any type of congestion
or back-pressure. So if there is congestion on the network,
HW can generate PAUSE frames to limit the traffic and packets
can be dropped in the DPDK.

So in this patch, removing the pause property from switch and
ethernet ports.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>