IMXBSP-79: drm/imx/dcss: fix tiled to linear plane transitions
commit  11977ec8b6b665097ae2f04eb6e74d1fd1e644c3 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When a plane transitions from a video tiled format (G1/G2) to linear,
without performing a disable/enable sequence, the screen will go
green because the scaler for that specific channel FREEZES.

The problem happens because we don't flag the DTRC channel as 'not
running' when we switch it to bypass mode. And also because, when DPR
is set up, it checks to see whether the DTRC channel is running or not
in order to properly set the addresses. But, since it finds the channel
running, it continues to use the internal DTRC ARIDs, instead of
using the actual buffer location in DDR. Hence DPR fails with an AXI
read error, which leads to scaler freezing.

The solution is to properly flag the DTRC channel when switching to
bypass mode and move the dcss_dpr_addr_set() after the
dcss_plane_set_dtrc_base() is called, so that the DTRC channel is
already flagged as "not running" by the time we set the DPR addresses.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>