media: platform: microchip: Add On-Screen Display support to mchp-dscmi
commit 5fa0993bb64ab87ac61ee7003dfdbdc6e9df1677 from
https://github.com/linux4microchip/linux.git

The On-Screen Display feature (OSD) in can be an Image or Video.
The raw image is compressed as H264 frame and an overlay image is
added on top of this H264 frame.

The OSD also displays this compression ratio for the last 10 frames
value along with the overlay image.

The X position (horizontal) and Y position (vertical) of the overlay
image can be altered using av4l2 command.
Similarly, the overlay also can be enabled or disabled using a v4l2
command.

Signed-off-by: Vattipalli Praveen <praveen.kumar@microchip.com>
[Conor: fixed up commit message, comments, whitespace & a build error]
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>