LF-7981-8: drm: bridge: it6161: Use special horizontal flags
commit  cf7d0b1b43391363449062066eb6ea5d9d0e4ddc from
https://github.com/nxp-imx/linux-imx lf-6.18.y

This IP does not work when the horizontal blaking area is in pixels, but
works when this area is in bytes.
Add the flags for disabling the horizontal blanking area, so that the
DSI host can take care of this and send the blankings in bytes instead
of pixels.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>