watchdog: sbsa_gwdt: Apply the Errata workaround seen on CN10K Processors
commit 1c97de5e7c9d304334fe561d510e259509147944 from
git@git.assembla.com:cavium/WindRiver.linux.git

[ Ported from 6.1 branch ]

As per the Errata IPBUGTI-37100 on Marvell's CN10K series of processors,
the watchdog time is not calculated based on arch_timer_get_cntfrq().
Instead, It is based on the Internal SCLK. In certain  frequency settings
the clkrate is faster leading to an early watchdog  timeout.

To avoid this always read the clk frequency provided by the firmware
in the watchdog device node. In case of system's without the errata
the arch_timer_get_cntfrq() will be used to find the clk frequency.

Change-Id: I29cc8047090592d718810b276e708b05b0728fd6
Signed-off-by: George Cherian <george.cherian@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/101203
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>