i2c: designware: introduce a custom scl recovery for SoCFPGA platforms
commit 73b66146e5a976b1c4b3fe3a14b950b54fb080eb from
https://github.com/altera-opensource/linux-socfpga.git

[upstream commit b2f670af1feab3325cce5a251128f30711148771]

The I2C pins on the SoCFPGA platforms do not go through a GPIO module,
thus cannot be recovered by the default method of by doing a GPIO access.
Only a reset of the I2C IP block can a recovery be successful, so this
change effectively resets the I2C controller, NOT any attached clients.

Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Dinh Nguyen <dinguyen@kernel.org>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>