soc: tegra: Add config variable for platform helper support
commit ac736e514f430513951a2567f98356ff25d4d7dd from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git jetson_36.3

BugLink: https://bugs.launchpad.net/bugs/2028400

Platform helper support APIs are available on some of the
core kernel and these APIs are used by out-of-tree modules.
To find out from out-of-tree modules, whether core kernel
support the platform helpers or not, add the CONFIG variables.

The out-of-tree modules will check for this config and then
include/use the APIs provided by the platform helper functions.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: Vedashree Vidwans <vvidwans@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>