rpmsg: imx_rpmsg: Change the notifying to timeout mode
commit  d31dd73798ed91695aeb430d966418e9e6e7bf92 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Currently, after the rpmsg link up it will use mailbox
channel in blocking mode to send the notifications. So
it will be stuck in the notification when the remote side
unable to respond for some reason, such as crashed.

To avoid interlock, this patch changes the notifying to
timeout mode, and treats the virtqueue as broken when
the remote side doesn't respond within 1 second.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>