drivers: iio: adc: xilinx-ams: Add over temperature interrupts
commit 996f8822f7481052efd603133fe768d7b7836bd7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

This code adds the functionality to cause and handle an interrupt when an
over temperature event occurs.

The interrupt can be enabled/disabled and the threshold for alarm
can be set through sysfs (as with the other temperature interrupts).

Signed-off-by: Conall O'Griofa <conall.ogriofa@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
[michals:
v6.12 update
iio: adc: xilinx-ams: Make compilable for 32 bit architectures
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>