media: platform: microchip: Fix OSD boundary exceeding the frame boundary
commit f0bc6b0ee0c857741ed781710ff2bf1c14026eed from
https://github.com/linux4microchip/linux.git

Using the v4l2 command the OSD feature can be disabled, but actually it is
not disabled in the hardware, instead it is shifted to out of resolution
boundary(frame boundary), so that it won't appear on the screen. But when
the resoution is increased, the OSD reappears on the screen even though
it is disabled.

To fix this a separate register was added to the IP, so that the OSD will
get disabled in the hardware itself.

Signed-off-by: Vattipalli Praveen <praveen.kumar@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>